Add files via upload
This script is designed to find and manage inactive users in an Azure Active Directory environment. It begins by connecting to Azure AD using the Connect-AzureAD cmdlet. The number of days since the last logon is set to 90 and stored in the $DaysInactive variable. The $InactiveDate variable is then set to the current date minus the number of days stored in $DaysInactive.

It then finds inactive users by using the Get-AzureADUser cmdlet to retrieve all users and filtering them by the LastDirSyncTime and AccountEnabled attributes. The script then exports the results to a CSV file using the Export-Csv cmdlet.

Finally, it disables all inactive users by looping through the list of users and using the Set-AzureADUser cmdlet to set the AccountEnabled attribute to $false for each user.